+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-03-22
+
+### Added
+
+- `.epub` input.
+
 ## [0.3.0] - 2026-02-26
 
 ### Added
@@ -18,9 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- .pdf input (straightforward).
+- `.pdf` input (straightforward).
 
 ## [0.1.0] - 2026-02-19
 
-Takes and returns .docx file.  
+Takes and returns `.docx` file.  
 Very first simpliest version.